updating the admin page to table
