+++ src/App.css
@@ -62,5 +62,5 @@
   display: flex;
   flex-direction: row;
   flex-wrap: wrap;
-  /* justify-content: center; */
+  justify-content: center;
 }
